Add event and bootable trait support to controller
